requirements.txt
alembic==1.5.8
autopep8==1.5.4
bcrypt==3.2.0
blinker==1.4
boto3==1.16.19
botocore==1.19.19
cffi==1.14.2
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
limits==1.5.1
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyStemmer==2.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
